2022Wechat/requirements.txt
torch-ema==0.3
torchaudio==0.7.2
torchmetrics==0.6.0
torchvision==0.8.2+cu110
pytorch-lightning==1.5.10
optuna==2.10.0
overrides==6.1.0
boto3==1.24.2
transformers==4.19.1
wandb==0.12.7
pandas==1.1.5
matplotlib==3.4.3
scikit-learn==1.0.2
tqdm==4.61.1